feat: refactor SignInPage to use SigninClient component and update Tailwind dark mode configuration
